Camera movement without collisions.
Input is separated from Camera class for lisibility.
Camera update its position and its direction.
It can return a view mat4 for glsl shader.

+++ include/Time.hpp
@@ -16,7 +16,7 @@ public:
 
 	bool			update(void);
 
-	unsigned int	delta_time(void);
+	float			delta_time(void);
 	unsigned int	fps(void);
 };
